get overdue borrows method created

+++ src/app/modules/Book/book.service.ts
@@ -1,5 +1,7 @@
 import { Book } from "@prisma/client";
 import prisma from "../../utils/prisma";
+import AppError from "../../utils/AppError";
+import httpStatus from "http-status";
 
 const createBookIntoDB = async (data: Book) => {
   const result = await prisma.book.create({
@@ -21,10 +23,20 @@ const getBookByIdFromDB = async (bookId: string) => {
     },
   });
 
+  if (!result) throw new AppError(httpStatus.NOT_FOUND, "Book not found");
+
   return result;
 };
 
 const updateBookIntoDB = async (bookId: string, data: Partial<Book>) => {
+  const book = await prisma.book.findUnique({
+    where: {
+      bookId,
+    },
+  });
+
+  if (!book) throw new AppError(httpStatus.NOT_FOUND, "Book not found");
+
   const result = await prisma.book.update({
     where: {
       bookId,
@@ -36,6 +48,14 @@ const updateBookIntoDB = async (bookId: string, data: Partial<Book>) => {
 };
 
 const deleteBookFromDB = async (bookId: string) => {
+  const book = await prisma.book.findUnique({
+    where: {
+      bookId,
+    },
+  });
+
+  if (!book) throw new AppError(httpStatus.NOT_FOUND, "Book not found");
+
   const result = await prisma.book.delete({
     where: {
       bookId,
