+++ src/app/modules/Member/member.service.ts
@@ -1,5 +1,7 @@
 import { Member } from "@prisma/client";
 import prisma from "../../utils/prisma";
+import AppError from "../../utils/AppError";
+import httpStatus from "http-status";
 
 const createMemberIntoDB = async (data: Member) => {
   const result = await prisma.member.create({
@@ -22,10 +24,20 @@ const getMemberByIdFromDB = async (memberId: string) => {
     },
   });
 
+  if (!result) throw new AppError(httpStatus.NOT_FOUND, "Member not found");
+
   return result;
 };
 
 const updateMemberIntoDB = async (memberId: string, data: Partial<Member>) => {
+  const member = await prisma.member.findUnique({
+    where: {
+      memberId,
+    },
+  });
+
+  if (!member) throw new AppError(httpStatus.NOT_FOUND, "Member not found");
+
   const result = await prisma.member.update({
     where: {
       memberId,
@@ -37,6 +49,14 @@ const updateMemberIntoDB = async (memberId: string, data: Partial<Member>) => {
 };
 
 const deleteMemberFromDB = async (memberId: string) => {
+  const member = await prisma.member.findUnique({
+    where: {
+      memberId,
+    },
+  });
+
+  if (!member) throw new AppError(httpStatus.NOT_FOUND, "Member not found");
+
   const result = await prisma.member.delete({
     where: {
       memberId,
